feat(04-02): Combined-UI Upload-Handler implementieren
- ChunkStream.h include hinzugefuegt
- Globale Variablen fuer Combined-Update-State (combinedUiError, combinedFwError, etc.)
- uiExtractTask: FreeRTOS-Task auf Core 0 mit tarGzStreamExpander und .bin-Exclude-Filter
- handleCombinedUiUpload: WebServer-Handler mit ChunkStream-Streaming, WDT-Deaktivierung
- VERSION.txt-Lesen und Verteilung auf firmware-version.txt + ui-version.txt (OTA-03)
- UPLOAD_FILE_ABORTED-Handling fuer sauberen State-Reset (OTA-04)
- Route /update/combined-ui registriert
- Rule1-Fix: ChunkStream min() Typ-Mismatch (volatile size_t) behoben

+++ firmware/src/ChunkStream.h
@@ -107,7 +107,7 @@ public:
                 if (_available == 0 && _eof) break;
                 if (_available == 0) break; // Timeout
             }
-            size_t canRead = min(len - totalRead, _available);
+            size_t canRead = min(len - totalRead, (size_t)_available);
             for (size_t i = 0; i < canRead; i++) {
                 buf[totalRead + i] = _buffer[_readPos];
                 _readPos = (_readPos + 1) % _bufSize;
